テストコードを別ファイルに分離し、重複を整理して圧縮
各モジュール内に埋め込まれていた #[cfg(test)] mod tests を
#[path] 属性で隣接する *_tests.rs に切り出し、プロダクトコードと
テストコードを分離。あわせてテスト自体も整理:

- config: ほぼ同一の TOML を繰り返していた watch フィルタ検証
  (排他・構文チェック) と必須フィールド検証をテーブル駆動化
  (69 テスト ≒1490 行 → 14 テスト ≒540 行)
- placeholder: プレースホルダ単体展開 11 テストをテーブル駆動で 1 本に
- router: マッチ/非マッチで対になっていたテストを統合 (27 → 11)
- actions: copy/move/command/execute/common で重複していた
  make_sink / write_file / ActionConfig 生成を test_support に共通化
- ハッシュ比較を再確認するだけで本来の対象を検証していなかった
  テスト 2 本 (common/move) を削除、未使用の cfg(test) ヘルパー
  validate_action_placeholders を削除

テスト対象の挙動カバレッジは維持 (cargo test 全パス)。
Cargo.lock は cat-watcher 1.2.1 への追従 (cargo が自動同期)。
