+++ cat-watcher/src/actions/command.rs
@@ -93,129 +93,5 @@ fn build_shell_command(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::config::{ActionType, LogRotation};
-    use crate::logger::Logger;
-    use std::sync::Arc;
-    use tempfile::tempdir;
-
-    fn make_action(shell: &str, command: &str, working_dir: &str) -> ActionConfig {
-        ActionConfig {
-            type_: ActionType::Command,
-            destination: None,
-            overwrite: None,
-            preserve_structure: None,
-            verify_integrity: None,
-            shell: Some(shell.to_string()),
-            command: Some(command.to_string()),
-            working_dir: Some(working_dir.to_string()),
-            program: None,
-            args: None,
-            message: None,
-        }
-    }
-
-    fn make_ctx(src: &std::path::Path, watch: &std::path::Path) -> PlaceholderContext {
-        PlaceholderContext::new(src, watch, "")
-    }
-
-    fn make_sink() -> ActionSink {
-        let dir = tempdir().unwrap();
-        let (logger, _) = Logger::for_action(
-            dir.path().to_str().unwrap().to_string(),
-            "test.log".to_string(),
-            LogRotation::Never,
-        )
-        .unwrap();
-        std::mem::forget(dir);
-        ActionSink::new(Arc::new(logger), None)
-    }
-
-    #[tokio::test]
-    async fn unknown_shell_returns_error() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("a.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("zsh", "echo hi", "");
-        let result = execute(&action, &ctx, &make_sink(), (1, 1)).await;
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("不明なシェル"));
-    }
-
-    #[cfg(not(windows))]
-    #[tokio::test]
-    async fn bash_spawns_successfully() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("a.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("bash", "echo hello", "");
-        assert!(execute(&action, &ctx, &make_sink(), (1, 1)).await.is_ok());
-    }
-
-    #[cfg(not(windows))]
-    #[tokio::test]
-    async fn sh_spawns_successfully() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("a.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("sh", "echo hello", "");
-        assert!(execute(&action, &ctx, &make_sink(), (1, 1)).await.is_ok());
-    }
-
-    #[cfg(target_os = "windows")]
-    #[tokio::test]
-    async fn cmd_spawns_successfully() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("a.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("cmd", "echo hello", "");
-        assert!(execute(&action, &ctx, &make_sink(), (1, 1)).await.is_ok());
-    }
-
-    #[cfg(target_os = "windows")]
-    #[tokio::test]
-    async fn placeholder_expands_in_command() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("report.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("cmd", "echo {Name}", "");
-        assert!(execute(&action, &ctx, &make_sink(), (1, 1)).await.is_ok());
-    }
-
-    #[cfg(target_os = "windows")]
-    #[tokio::test]
-    async fn working_dir_is_set() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("a.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("cmd", "echo hi", dir.path().to_str().unwrap());
-        assert!(execute(&action, &ctx, &make_sink(), (1, 1)).await.is_ok());
-    }
-
-    #[test]
-    fn build_shell_command_cmd() {
-        assert!(build_shell_command("cmd", "echo test").is_ok());
-    }
-
-    #[test]
-    fn build_shell_command_powershell() {
-        assert!(build_shell_command("powershell", "Get-Date").is_ok());
-    }
-
-    #[test]
-    fn build_shell_command_pwsh() {
-        assert!(build_shell_command("pwsh", "Get-Date").is_ok());
-    }
-
-    #[test]
-    fn build_shell_command_unknown() {
-        assert!(build_shell_command("zsh", "echo hi").is_err());
-    }
-}
+#[path = "command_tests.rs"]
+mod tests;
