+++ cat-watcher/src/actions/common.rs
@@ -93,104 +93,5 @@ pub async fn walk_files(src_dir: &Path) -> Result<Vec<PathBuf>, AppError> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::io::Write;
-    use tempfile::tempdir;
-
-    fn write_file(path: &Path, body: &[u8]) {
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent).unwrap();
-        }
-        std::fs::File::create(path).unwrap().write_all(body).unwrap();
-    }
-
-    #[tokio::test]
-    async fn hash_consistent() {
-        let dir = tempdir().unwrap();
-        let p = dir.path().join("x.bin");
-        write_file(&p, b"abcdef");
-        let h1 = hash_file_blake3(&p).await.unwrap();
-        let h2 = hash_file_blake3(&p).await.unwrap();
-        assert_eq!(h1, h2);
-    }
-
-    #[tokio::test]
-    async fn hash_differs_for_different_content() {
-        let dir = tempdir().unwrap();
-        let a = dir.path().join("a.bin");
-        let b = dir.path().join("b.bin");
-        write_file(&a, b"aaa");
-        write_file(&b, b"bbb");
-        assert_ne!(
-            hash_file_blake3(&a).await.unwrap(),
-            hash_file_blake3(&b).await.unwrap()
-        );
-    }
-
-    #[tokio::test]
-    async fn try_copy_once_copies_file() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("src.txt");
-        let dest = dir.path().join("dest.txt");
-        write_file(&src, b"hello");
-        try_copy_once(&src, &dest, false).await.unwrap();
-        assert_eq!(std::fs::read(&dest).unwrap(), b"hello");
-    }
-
-    #[tokio::test]
-    async fn try_copy_once_verify_integrity_ok() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("src.txt");
-        let dest = dir.path().join("dest.txt");
-        write_file(&src, b"payload");
-        try_copy_once(&src, &dest, true).await.unwrap();
-        assert!(dest.exists());
-    }
-
-    #[tokio::test]
-    async fn try_copy_once_integrity_mismatch_returns_error() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("src.txt");
-        let dest = dir.path().join("dest.txt");
-        write_file(&src, b"original");
-        write_file(&dest, b"tampered");
-        // dest は既存だが tokio::fs::copy で上書きされてしまうためハッシュは一致する。
-        // 不一致テスト: src を書き換えてから verify する
-        try_copy_once(&src, &dest, false).await.unwrap();
-        // dest 内容を上書きして不一致状態を作り hash 比較だけ確認
-        write_file(&dest, b"corrupted");
-        let src_hash = hash_file_blake3(&src).await.unwrap();
-        let dest_hash = hash_file_blake3(&dest).await.unwrap();
-        assert_ne!(src_hash, dest_hash);
-    }
-
-    #[test]
-    fn resolve_dest_path_flat() {
-        let src = Path::new("/watch/sub/a.txt");
-        let dest_root = Path::new("/dest");
-        let watch = Path::new("/watch");
-        let result = resolve_dest_path(src, dest_root, watch, false).unwrap();
-        assert_eq!(result, PathBuf::from("/dest/a.txt"));
-    }
-
-    #[test]
-    fn resolve_dest_path_preserve() {
-        let src = Path::new("/watch/sub/a.txt");
-        let dest_root = Path::new("/dest");
-        let watch = Path::new("/watch");
-        let result = resolve_dest_path(src, dest_root, watch, true).unwrap();
-        assert_eq!(result, PathBuf::from("/dest/sub/a.txt"));
-    }
-
-    #[tokio::test]
-    async fn walk_files_returns_all_files() {
-        let dir = tempdir().unwrap();
-        write_file(&dir.path().join("a.txt"), b"a");
-        write_file(&dir.path().join("sub/b.txt"), b"b");
-        write_file(&dir.path().join("sub/deep/c.txt"), b"c");
-        let mut files = walk_files(dir.path()).await.unwrap();
-        files.sort();
-        assert_eq!(files.len(), 3);
-    }
-}
+#[path = "common_tests.rs"]
+mod tests;
