+++ cat-watcher/src/actions/copy.rs
@@ -159,203 +159,5 @@ async fn copy_directory_recursive(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::config::{ActionType, LogRotation};
-    use crate::logger::Logger;
-    use std::io::Write;
-    use std::sync::Arc;
-    use tempfile::tempdir;
-
-    fn make_retry(count: u32) -> RetryConfig {
-        RetryConfig { count, interval_ms: 10 }
-    }
-
-    fn make_copy_action(
-        dest: &str,
-        overwrite: bool,
-        preserve_structure: bool,
-        verify_integrity: bool,
-    ) -> ActionConfig {
-        ActionConfig {
-            type_: ActionType::Copy,
-            destination: Some(dest.to_string()),
-            overwrite: Some(overwrite),
-            preserve_structure: Some(preserve_structure),
-            verify_integrity: Some(verify_integrity),
-            working_dir: None,
-            shell: None,
-            command: None,
-            program: None,
-            args: None,
-            message: None,
-        }
-    }
-
-    fn write_file(path: &Path, body: &[u8]) {
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent).unwrap();
-        }
-        let mut f = std::fs::File::create(path).unwrap();
-        f.write_all(body).unwrap();
-    }
-
-    fn make_sink() -> ActionSink {
-        let dir = tempdir().unwrap();
-        let (logger, _) = Logger::for_action(
-            dir.path().to_str().unwrap().to_string(),
-            "test.log".to_string(),
-            LogRotation::Never,
-        )
-        .unwrap();
-        std::mem::forget(dir);
-        ActionSink::new(Arc::new(logger), None)
-    }
-
-    #[tokio::test]
-    async fn copies_single_file_flat() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"hello");
-
-        let action = make_copy_action(dest.path().to_str().unwrap(), false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        let dest_file = dest.path().join("a.txt");
-        assert!(dest_file.exists());
-        assert_eq!(result, Some(dest_file));
-    }
-
-    #[tokio::test]
-    async fn preserves_subdir_structure() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("sub/deep/a.txt");
-        write_file(&src, b"hello");
-
-        let action = make_copy_action(dest.path().to_str().unwrap(), false, true, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert!(dest.path().join("sub/deep/a.txt").exists());
-    }
-
-    #[tokio::test]
-    async fn skips_when_overwrite_false_and_exists() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"new");
-        let dest_file = dest.path().join("a.txt");
-        write_file(&dest_file, b"old");
-
-        let action = make_copy_action(dest.path().to_str().unwrap(), false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert_eq!(result, None);
-        assert_eq!(std::fs::read(&dest_file).unwrap(), b"old");
-    }
-
-    #[tokio::test]
-    async fn overwrites_when_overwrite_true() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"new");
-        let dest_file = dest.path().join("a.txt");
-        write_file(&dest_file, b"old");
-
-        let action = make_copy_action(dest.path().to_str().unwrap(), true, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert_eq!(std::fs::read(&dest_file).unwrap(), b"new");
-    }
-
-    #[tokio::test]
-    async fn verify_integrity_passes_for_identical_content() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"some payload to hash");
-
-        let action = make_copy_action(dest.path().to_str().unwrap(), false, false, true);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert!(result.is_some());
-        assert!(dest.path().join("a.txt").exists());
-    }
-
-    #[tokio::test]
-    async fn copies_directory_recursively() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src_dir = watch.path().join("mydir");
-        write_file(&src_dir.join("a.txt"), b"a");
-        write_file(&src_dir.join("sub/b.txt"), b"b");
-
-        let action = make_copy_action(dest.path().to_str().unwrap(), false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src_dir, watch.path(), "");
-
-        let result = execute(&action, &src_dir, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert_eq!(result, Some(dest.path().join("mydir")));
-        assert!(dest.path().join("mydir/a.txt").exists());
-        assert!(dest.path().join("mydir/sub/b.txt").exists());
-    }
-
-    #[tokio::test]
-    async fn destination_with_multiple_placeholders_creates_intermediate_dirs() {
-        let watch = tempdir().unwrap();
-        let dest_root = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"hello");
-
-        let dest_template = format!(
-            "{}/{{Date}}/TESTDATA/{{Time}}",
-            dest_root.path().to_str().unwrap()
-        );
-        let action = make_copy_action(&dest_template, false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        let dest_path = result.expect("コピー成功");
-
-        assert!(dest_path.exists(), "コピー先ファイルが存在しない: {}", dest_path.display());
-        assert_eq!(dest_path.file_name().unwrap(), "a.txt");
-        let parent = dest_path.parent().unwrap();
-        assert_eq!(parent.file_name().unwrap().to_str().unwrap().len(), 6);
-        let grandparent = parent.parent().unwrap();
-        assert_eq!(grandparent.file_name().unwrap(), "TESTDATA");
-        let great_grandparent = grandparent.parent().unwrap();
-        assert_eq!(great_grandparent.file_name().unwrap().to_str().unwrap().len(), 8);
-    }
-
-    #[tokio::test]
-    async fn destination_placeholder_expands_in_dest() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"hello");
-
-        let dest_template = format!("{}/{{BaseName}}", dest.path().to_str().unwrap());
-        let action = make_copy_action(&dest_template, false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        let expected = dest.path().join("a").join("a.txt");
-        assert_eq!(result.as_deref(), Some(expected.as_path()));
-        assert!(expected.exists());
-    }
-}
+#[path = "copy_tests.rs"]
+mod tests;
