+++ cat-watcher/src/actions/execute.rs
@@ -52,86 +52,5 @@ pub async fn execute(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::config::{ActionType, LogRotation};
-    use crate::logger::Logger;
-    use std::sync::Arc;
-    use tempfile::tempdir;
-
-    fn make_action(program: &str, args: Vec<&str>, working_dir: &str) -> ActionConfig {
-        ActionConfig {
-            type_: ActionType::Execute,
-            destination: None,
-            overwrite: None,
-            preserve_structure: None,
-            verify_integrity: None,
-            shell: None,
-            command: None,
-            working_dir: Some(working_dir.to_string()),
-            program: Some(program.to_string()),
-            args: Some(args.into_iter().map(|s| s.to_string()).collect()),
-            message: None,
-        }
-    }
-
-    fn make_ctx(src: &std::path::Path, watch: &std::path::Path) -> PlaceholderContext {
-        PlaceholderContext::new(src, watch, "")
-    }
-
-    fn make_sink() -> ActionSink {
-        let dir = tempdir().unwrap();
-        let (logger, _) = Logger::for_action(
-            dir.path().to_str().unwrap().to_string(),
-            "test.log".to_string(),
-            LogRotation::Never,
-        )
-        .unwrap();
-        std::mem::forget(dir);
-        ActionSink::new(Arc::new(logger), None)
-    }
-
-    #[cfg(target_os = "windows")]
-    #[tokio::test]
-    async fn spawns_program_successfully() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("a.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("cmd.exe", vec!["/C", "echo hello"], "");
-        assert!(execute(&action, &ctx, &make_sink(), (1, 1)).await.is_ok());
-    }
-
-    #[cfg(target_os = "windows")]
-    #[tokio::test]
-    async fn placeholder_expands_in_args() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("report.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("cmd.exe", vec!["/C", "echo {FullName}"], "");
-        assert!(execute(&action, &ctx, &make_sink(), (1, 1)).await.is_ok());
-    }
-
-    #[cfg(target_os = "windows")]
-    #[tokio::test]
-    async fn working_dir_is_set() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("a.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("cmd.exe", vec!["/C", "echo hi"], dir.path().to_str().unwrap());
-        assert!(execute(&action, &ctx, &make_sink(), (1, 1)).await.is_ok());
-    }
-
-    #[tokio::test]
-    async fn nonexistent_program_returns_error() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("a.txt");
-        std::fs::write(&src, b"x").unwrap();
-        let ctx = make_ctx(&src, dir.path());
-        let action = make_action("nonexistent_program_xyz.exe", vec![], "");
-        let result = execute(&action, &ctx, &make_sink(), (1, 1)).await;
-        assert!(result.is_err());
-    }
-}
+#[path = "execute_tests.rs"]
+mod tests;
