+++ cat-watcher/src/actions/move.rs
@@ -220,209 +220,5 @@ fn windows_is_cross_device(_e: &std::io::Error) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use super::super::common::hash_file_blake3;
-    use crate::config::{ActionType, LogRotation};
-    use crate::logger::Logger;
-    use std::io::Write;
-    use std::sync::Arc;
-    use tempfile::tempdir;
-
-    fn make_retry(count: u32) -> RetryConfig {
-        RetryConfig { count, interval_ms: 10 }
-    }
-
-    fn make_move_action(
-        dest: &str,
-        overwrite: bool,
-        preserve_structure: bool,
-        verify_integrity: bool,
-    ) -> ActionConfig {
-        ActionConfig {
-            type_: ActionType::Move,
-            destination: Some(dest.to_string()),
-            overwrite: Some(overwrite),
-            preserve_structure: Some(preserve_structure),
-            verify_integrity: Some(verify_integrity),
-            working_dir: None,
-            shell: None,
-            command: None,
-            program: None,
-            args: None,
-            message: None,
-        }
-    }
-
-    fn write_file(path: &Path, body: &[u8]) {
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent).unwrap();
-        }
-        std::fs::File::create(path).unwrap().write_all(body).unwrap();
-    }
-
-    fn make_sink() -> ActionSink {
-        let dir = tempdir().unwrap();
-        let (logger, _) = Logger::for_action(
-            dir.path().to_str().unwrap().to_string(),
-            "test.log".to_string(),
-            LogRotation::Never,
-        )
-        .unwrap();
-        std::mem::forget(dir);
-        ActionSink::new(Arc::new(logger), None)
-    }
-
-    #[tokio::test]
-    async fn moves_single_file() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"hello");
-
-        let action = make_move_action(dest.path().to_str().unwrap(), false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert_eq!(result, Some(dest.path().join("a.txt")));
-        assert!(dest.path().join("a.txt").exists());
-        assert!(!src.exists(), "元ファイルが残っている");
-    }
-
-    #[tokio::test]
-    async fn skips_when_overwrite_false_and_exists() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"new");
-        write_file(&dest.path().join("a.txt"), b"old");
-
-        let action = make_move_action(dest.path().to_str().unwrap(), false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert_eq!(result, None);
-        assert_eq!(std::fs::read(dest.path().join("a.txt")).unwrap(), b"old");
-        assert!(src.exists(), "スキップ時は元ファイルを保持");
-    }
-
-    #[tokio::test]
-    async fn overwrites_when_overwrite_true() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"new");
-        write_file(&dest.path().join("a.txt"), b"old");
-
-        let action = make_move_action(dest.path().to_str().unwrap(), true, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert_eq!(std::fs::read(dest.path().join("a.txt")).unwrap(), b"new");
-        assert!(!src.exists());
-    }
-
-    #[tokio::test]
-    async fn preserves_subdir_structure() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("sub/deep/a.txt");
-        write_file(&src, b"hello");
-
-        let action = make_move_action(dest.path().to_str().unwrap(), false, true, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert!(dest.path().join("sub/deep/a.txt").exists());
-        assert!(!src.exists());
-    }
-
-    #[tokio::test]
-    async fn moves_directory_recursively() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src_dir = watch.path().join("mydir");
-        write_file(&src_dir.join("a.txt"), b"a");
-        write_file(&src_dir.join("sub/b.txt"), b"b");
-
-        let action = make_move_action(dest.path().to_str().unwrap(), false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src_dir, watch.path(), "");
-
-        let result = execute(&action, &src_dir, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert_eq!(result, Some(dest.path().join("mydir")));
-        assert!(dest.path().join("mydir/a.txt").exists());
-        assert!(dest.path().join("mydir/sub/b.txt").exists());
-        assert!(!src_dir.exists(), "移動元フォルダが削除されていない");
-    }
-
-    #[tokio::test]
-    async fn verify_integrity_passes_on_same_volume() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"payload for hash");
-
-        let action = make_move_action(dest.path().to_str().unwrap(), false, false, true);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        assert!(result.is_some());
-        assert!(!src.exists());
-    }
-
-    #[tokio::test]
-    async fn destination_with_placeholder_expands() {
-        let watch = tempdir().unwrap();
-        let dest_root = tempdir().unwrap();
-        let src = watch.path().join("a.txt");
-        write_file(&src, b"hello");
-
-        let dest_template = format!(
-            "{}/{{Date}}/{{Time}}",
-            dest_root.path().to_str().unwrap()
-        );
-        let action = make_move_action(&dest_template, false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await.unwrap();
-        let dest_path = result.expect("移動成功");
-        assert!(dest_path.exists());
-        assert_eq!(dest_path.file_name().unwrap(), "a.txt");
-        assert!(!src.exists());
-    }
-
-    #[tokio::test]
-    async fn source_preserved_on_final_failure() {
-        let watch = tempdir().unwrap();
-        let dest = tempdir().unwrap();
-        let src = watch.path().join("nonexistent.txt");
-
-        let action = make_move_action(dest.path().to_str().unwrap(), false, false, false);
-        let retry = make_retry(0);
-        let ctx = PlaceholderContext::new(&src, watch.path(), "");
-
-        let result = execute(&action, &src, &ctx, &retry, &make_sink(), (1, 1)).await;
-        assert!(result.is_err(), "存在しないファイルの move はエラー");
-    }
-
-    #[tokio::test]
-    async fn hash_mismatch_on_fallback_removes_dest_not_src() {
-        let dir = tempdir().unwrap();
-        let src = dir.path().join("src.txt");
-        let dest = dir.path().join("dest.txt");
-        write_file(&src, b"original");
-        write_file(&dest, b"tampered");
-
-        let src_hash = hash_file_blake3(&src).await.unwrap();
-        let dest_hash = hash_file_blake3(&dest).await.unwrap();
-        assert_ne!(src_hash, dest_hash, "内容が異なればハッシュも異なる");
-        assert!(src.exists());
-    }
-}
+#[path = "move_tests.rs"]
+mod tests;
