+++ cat-watcher/src/config.rs
@@ -565,1495 +565,5 @@ fn collect_action_placeholder_errors(action: &ActionConfig, rule_name: &str, err
 }
 
 #[cfg(test)]
-fn validate_action(action: &ActionConfig, rule_name: &str) -> Result<(), AppError> {
-	let mut errors = Vec::new();
-	collect_action_errors(action, rule_name, &mut errors);
-	finish_validation(errors)
-}
-
-#[cfg(test)]
-fn validate_action_placeholders(action: &ActionConfig, rule_name: &str) -> Result<(), AppError> {
-	let mut errors = Vec::new();
-	collect_action_placeholder_errors(action, rule_name, &mut errors);
-	finish_validation(errors)
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use tempfile::tempdir;
-
-	fn sanitize_path(path: &std::path::Path) -> String {
-		path.to_str().unwrap().replace('\\', "/")
-	}
-
-	// =========================================================
-	// ヘルパー: バリデーションで watch.path / destination に
-	// 実在するディレクトリが必要なので、一時ディレクトリのパスを
-	// テンプレートに埋め込む
-	// =========================================================
-
-	fn make_rules_toml(watch_path: &str, action_block: &str) -> String {
-		let watch_path = watch_path.replace('\\', "/");
-		format!(r#"
-			[[rules]]
-			enabled = true
-			name = "test-rule"
-
-			[rules.watch]
-			path = "{watch_path}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			events = ["create", "modify"]
-
-			[[rules.actions]]
-			{action_block}
-		"#)
-	}
-
-	fn make_rules_toml_with_watch(watch_block: &str, watch_path: &str) -> String {
-		let watch_path = watch_path.replace('\\', "/");
-		format!(r#"
-			[[rules]]
-			enabled = true
-			name = "test-rule"
-
-			[rules.watch]
-			path = "{watch_path}"
-			{watch_block}
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hello"
-			working_dir = ""
-		"#)
-	}
-
-	// =========================================================
-	// GlobalConfig パーステスト
-	// =========================================================
-
-	fn make_global_toml(dir_path: &str) -> String {
-		format!(r#"
-			[retry]
-			count = 3
-			interval_ms = 1000
-
-			[system_log]
-			enabled = true
-			dir = "{dir_path}"
-			file_name = "system_{{Date}}.log"
-			rotation = "daily"
-			level = "info"
-			console = true
-		"#)
-	}
-
-	#[test]
-	fn test_parse_global_config() {
-		let dir = tempdir().unwrap();
-		let dir_path = sanitize_path(dir.path());
-		let config: GlobalConfig = toml::from_str(&make_global_toml(&dir_path)).unwrap();
-		assert_eq!(config.retry.count, 3);
-		assert_eq!(config.retry.interval_ms, 1000);
-		assert_eq!(config.system_log.dir, dir_path);
-		assert!(config.system_log.console);
-	}
-
-	#[test]
-	fn test_parse_global_config_all_log_levels() {
-		let dir = tempdir().unwrap();
-		let dir_path = sanitize_path(dir.path());
-		for level in &["trace", "debug", "info", "warn", "error"] {
-			let toml_str = format!(r#"
-				[retry]
-				count = 1
-				interval_ms = 500
-
-				[system_log]
-				dir = "{dir_path}"
-				file_name = "system.log"
-				rotation = "daily"
-				level = "{level}"
-			"#);
-			let result: Result<GlobalConfig, _> = toml::from_str(&toml_str);
-			assert!(result.is_ok(), "level '{}' のパースに失敗", level);
-		}
-	}
-
-	#[test]
-	fn test_parse_global_config_invalid_log_level() {
-		let dir = tempdir().unwrap();
-		let dir_path = sanitize_path(dir.path());
-		let toml_str = format!(r#"
-			[retry]
-			count = 1
-			interval_ms = 500
-
-			[system_log]
-			dir = "{dir_path}"
-			file_name = "system.log"
-			rotation = "daily"
-			level = "verbose"
-		"#);
-		let result: Result<GlobalConfig, _> = toml::from_str(&toml_str);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_parse_global_config_missing_section() {
-		// system_log セクション自体が無い場合はエラー
-		let toml_str = r#"
-			[retry]
-			count = 1
-			interval_ms = 500
-		"#;
-		let result: Result<GlobalConfig, _> = toml::from_str(toml_str);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_parse_global_config_rejects_legacy_global_section() {
-		// 旧 [global] 形式は deny_unknown_fields で明示エラーになる
-		let toml_str = r#"
-			[global]
-			log_level = "info"
-			log_dir = "logs"
-			log_file_name = "app.log"
-			log_rotation = "daily"
-			retry_count = 3
-			retry_interval_ms = 1000
-		"#;
-		let result: Result<GlobalConfig, _> = toml::from_str(toml_str);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_parse_global_config_rejects_unknown_key() {
-		// system_log に未知キー（旧 log_to_file 等）があればエラー
-		let dir = tempdir().unwrap();
-		let dir_path = sanitize_path(dir.path());
-		let toml_str = format!(r#"
-			[retry]
-			count = 1
-			interval_ms = 500
-
-			[system_log]
-			dir = "{dir_path}"
-			file_name = "system.log"
-			rotation = "daily"
-			level = "info"
-			log_to_file = true
-		"#);
-		let result: Result<GlobalConfig, _> = toml::from_str(&toml_str);
-		assert!(result.is_err());
-	}
-
-	// =========================================================
-	// GlobalConfig バリデーションテスト
-	// =========================================================
-
-	fn make_global(dir: &str, file_name: &str) -> GlobalConfig {
-		GlobalConfig {
-			retry: RetryConfig { count: 3, interval_ms: 1000 },
-			system_log: SystemLogConfig {
-				enabled: true,
-				dir: dir.to_string(),
-				file_name: file_name.to_string(),
-				rotation: LogRotation::Daily,
-				level: LogLevel::Info,
-				console: true,
-			},
-		}
-	}
-
-	#[test]
-	fn test_validate_global_config_empty_log_dir() {
-		let config = make_global("   ", "app.log");
-		assert!(validate_global_config(&config).is_err());
-	}
-
-	#[test]
-	fn test_validate_global_config_dir_not_exist() {
-		let config = make_global("nonexistent_dir_xyz_12345", "app.log");
-		assert!(validate_global_config(&config).is_err());
-	}
-
-	#[test]
-	fn test_validate_global_config_empty_file_name() {
-		let dir = tempdir().unwrap();
-		let config = make_global(dir.path().to_str().unwrap(), "  ");
-		assert!(validate_global_config(&config).is_err());
-	}
-
-	#[test]
-	fn test_validate_global_config_invalid_placeholder_in_file_name() {
-		let dir = tempdir().unwrap();
-		let config = make_global(dir.path().to_str().unwrap(), "app_{Name}.log");
-		assert!(validate_global_config(&config).is_err());
-	}
-
-	#[test]
-	fn test_validate_global_config_valid() {
-		let dir = tempdir().unwrap();
-		let config = make_global(dir.path().to_str().unwrap(), "app_{Date}.log");
-		assert!(validate_global_config(&config).is_ok());
-	}
-
-	// =========================================================
-	// RulesConfig パーステスト
-	// =========================================================
-
-	#[test]
-	fn test_parse_rules_config_copy_action() {
-		let dir = tempdir().unwrap();
-		let dest = tempdir().unwrap();
-		let toml_str = make_rules_toml(
-			&sanitize_path(dir.path()),
-			&format!(r#"
-				type = "copy"
-				destination = "{}"
-				overwrite = true
-				verify_integrity = true
-				preserve_structure = false
-			"#, sanitize_path(dest.path())),
-		);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		assert_eq!(config.rules.len(), 1);
-		assert_eq!(config.rules[0].name, "test-rule");
-	}
-
-	#[test]
-	fn test_parse_rules_config_move_action() {
-		let dir = tempdir().unwrap();
-		let dest = tempdir().unwrap();
-		let toml_str = make_rules_toml(
-			&sanitize_path(dir.path()),
-			&format!(r#"
-				type = "move"
-				destination = "{}"
-				overwrite = false
-				verify_integrity = false
-				preserve_structure = true
-			"#, sanitize_path(dest.path())),
-		);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		assert_eq!(config.rules[0].actions[0].overwrite, Some(false));
-	}
-
-	#[test]
-	fn test_parse_rules_config_command_action() {
-		let dir = tempdir().unwrap();
-		let toml_str = make_rules_toml(
-			&sanitize_path(dir.path()),
-			r#"
-				type = "command"
-				shell = "cmd"
-				command = "echo hello"
-				working_dir = ""
-			"#,
-		);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		assert_eq!(config.rules[0].actions[0].command, Some("echo hello".to_string()));
-	}
-
-	#[test]
-	fn test_parse_rules_config_execute_action() {
-		let dir = tempdir().unwrap();
-		let toml_str = make_rules_toml(
-			&sanitize_path(dir.path()),
-			r#"
-				type = "execute"
-				program = "notepad.exe"
-				args = ["{FullName}"]
-				working_dir = ""
-			"#,
-		);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		assert_eq!(config.rules[0].actions[0].program, Some("notepad.exe".to_string()));
-	}
-
-	#[test]
-	fn test_parse_rules_config_with_regex_instead_of_patterns() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "regex-rule"
-
-			[rules.watch]
-			path = "{}"
-			recursive = false
-			target = "directory"
-			include_hidden = true
-			regex = "^report_\\d+\\.csv$"
-			exclude_patterns = ["*.tmp"]
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "powershell"
-			command = "echo test"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		assert!(config.rules[0].watch.patterns.is_none());
-		assert!(config.rules[0].watch.regex.is_some());
-	}
-
-	#[test]
-	fn test_parse_rules_config_all_events() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = false
-			name = "all-events"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "both"
-			include_hidden = false
-			patterns = ["*"]
-			exclude_patterns = []
-			events = ["create", "modify", "delete", "rename"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo done"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		assert_eq!(config.rules[0].watch.events.len(), 4);
-	}
-
-	#[test]
-	fn test_parse_rules_config_invalid_action_type() {
-		let dir = tempdir().unwrap();
-		let toml_str = make_rules_toml(
-			&sanitize_path(dir.path()),
-			r#"type = "delete""#,
-		);
-		let result: Result<RulesConfig, _> = toml::from_str(&toml_str);
-		assert!(result.is_err());
-	}
-
-	// =========================================================
-	// RulesConfig バリデーションテスト
-	// =========================================================
-
-	#[test]
-	fn test_validate_rules_empty_rules() {
-		let config = RulesConfig { rules: vec![] };
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_rules_empty_name() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "  "
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_rules_empty_actions() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "no-actions"
-			actions = []
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			events = ["create"]
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_rules_empty_events() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "no-events"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			events = []
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	// =========================================================
-	// patterns / regex 排他チェック
-	// =========================================================
-
-	#[test]
-	fn test_validate_rules_patterns_and_regex_both_present() {
-		let dir = tempdir().unwrap();
-		let toml_str = make_rules_toml_with_watch(
-			r#"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			regex = "^test"
-			exclude_patterns = []
-			events = ["create"]
-			"#,
-			dir.path().to_str().unwrap(),
-		);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_rules_patterns_and_regex_both_absent() {
-		let dir = tempdir().unwrap();
-		let toml_str = make_rules_toml_with_watch(
-			r#"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			exclude_patterns = []
-			events = ["create"]
-			"#,
-			dir.path().to_str().unwrap(),
-		);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	// =========================================================
-	// watch.path 存在チェック
-	// =========================================================
-
-	#[test]
-	fn test_validate_rules_watch_path_not_exist() {
-		let toml_str = make_rules_toml(
-			"C:/nonexistent_path_xyz_12345",
-			r#"
-				type = "command"
-				shell = "cmd"
-				command = "echo hi"
-				working_dir = ""
-			"#,
-		);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	// =========================================================
-	// glob / regex 構文チェック
-	// =========================================================
-
-	#[test]
-	fn test_validate_rules_invalid_glob_pattern() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "bad-glob"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["[invalid"]
-			exclude_patterns = []
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_rules_invalid_regex() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "bad-regex"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			regex = "(unclosed"
-			exclude_patterns = []
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_rules_invalid_exclude_glob() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "bad-exclude"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = ["[bad"]
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	// =========================================================
-	// exclude_regex / exclude_dir_patterns / exclude_dir_regex (#28)
-	// =========================================================
-
-	#[test]
-	fn test_validate_exclude_patterns_and_exclude_regex_both_set() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "both-exclude"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = ["*.tmp"]
-			exclude_regex = "^debug"
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_exclude_regex_only_valid() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "exclude-regex-only"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			exclude_regex = "^debug_\\d+"
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_ok());
-	}
-
-	#[test]
-	fn test_validate_invalid_exclude_regex() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "bad-exclude-regex"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			exclude_regex = "(unclosed"
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_exclude_dir_patterns_and_exclude_dir_regex_both_set() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "both-dir-exclude"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			exclude_dir_patterns = ["node_modules"]
-			exclude_dir_regex = "^\\."
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_exclude_dir_patterns_valid() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "dir-patterns-ok"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			exclude_dir_patterns = ["node_modules", ".git"]
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_ok());
-	}
-
-	#[test]
-	fn test_validate_invalid_exclude_dir_glob() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "bad-dir-glob"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			exclude_dir_patterns = ["[bad"]
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_invalid_exclude_dir_regex() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "bad-dir-regex"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			exclude_dir_regex = "(unclosed"
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_exclude_dir_regex_valid() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "dir-regex-ok"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			exclude_dir_regex = "^\\."
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_ok());
-	}
-
-	// =========================================================
-	// dir_patterns / dir_regex (#28)
-	// =========================================================
-
-	#[test]
-	fn test_validate_dir_patterns_and_dir_regex_both_set() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "both-dir-include"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			dir_patterns = ["src"]
-			dir_regex = "^reports"
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_dir_patterns_valid() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "dir-patterns-include-ok"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			dir_patterns = ["src", "reports"]
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_ok());
-	}
-
-	#[test]
-	fn test_validate_invalid_dir_glob() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "bad-dir-include-glob"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			dir_patterns = ["[bad"]
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	#[test]
-	fn test_validate_dir_regex_valid() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "dir-regex-include-ok"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			dir_regex = "^reports_\\d+"
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_ok());
-	}
-
-	#[test]
-	fn test_validate_invalid_dir_regex() {
-		let dir = tempdir().unwrap();
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "bad-dir-regex-include"
-
-			[rules.watch]
-			path = "{}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			dir_regex = "(unclosed"
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo hi"
-			working_dir = ""
-		"#, sanitize_path(dir.path()));
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_err());
-	}
-
-	// =========================================================
-	// validate_action: Copy / Move
-	// =========================================================
-
-	#[test]
-	fn test_validate_action_copy_valid() {
-		let dest = tempdir().unwrap();
-		let action = ActionConfig {
-			type_: ActionType::Copy,
-			destination: Some(dest.path().to_str().unwrap().to_string()),
-			overwrite: Some(true),
-			verify_integrity: Some(true),
-			preserve_structure: Some(false),
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_ok());
-	}
-
-	#[test]
-	fn test_validate_action_copy_missing_destination() {
-		let action = ActionConfig {
-			type_: ActionType::Copy,
-			destination: None,
-			overwrite: Some(true),
-			verify_integrity: Some(true),
-			preserve_structure: Some(false),
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_validate_action_copy_missing_overwrite() {
-		let dest = tempdir().unwrap();
-		let action = ActionConfig {
-			type_: ActionType::Copy,
-			destination: Some(dest.path().to_str().unwrap().to_string()),
-			overwrite: None,
-			verify_integrity: Some(true),
-			preserve_structure: Some(false),
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_validate_action_copy_missing_verify_integrity() {
-		let dest = tempdir().unwrap();
-		let action = ActionConfig {
-			type_: ActionType::Copy,
-			destination: Some(dest.path().to_str().unwrap().to_string()),
-			overwrite: Some(true),
-			verify_integrity: None,
-			preserve_structure: Some(false),
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_validate_action_copy_missing_preserve_structure() {
-		let dest = tempdir().unwrap();
-		let action = ActionConfig {
-			type_: ActionType::Copy,
-			destination: Some(dest.path().to_str().unwrap().to_string()),
-			overwrite: Some(true),
-			verify_integrity: Some(true),
-			preserve_structure: None,
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_validate_action_copy_destination_not_exist() {
-		let action = ActionConfig {
-			type_: ActionType::Copy,
-			destination: Some("C:/nonexistent_dest_xyz_99999".to_string()),
-			overwrite: Some(true),
-			verify_integrity: Some(true),
-			preserve_structure: Some(false),
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_validate_action_copy_destination_with_placeholder_ok() {
-		// destination 中の {Date} 以降はランタイム展開されるため、ルートだけ実在すれば OK
-		let dest_root = tempdir().unwrap();
-		let dest_template = format!("{}/{{Date}}/sub", sanitize_path(dest_root.path()));
-		let action = ActionConfig {
-			type_: ActionType::Copy,
-			destination: Some(dest_template),
-			overwrite: Some(true),
-			verify_integrity: Some(false),
-			preserve_structure: Some(false),
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_ok());
-	}
-
-	#[test]
-	fn test_validate_action_copy_destination_with_placeholder_root_missing() {
-		// プレースホルダー前のルートが存在しなければエラー
-		let action = ActionConfig {
-			type_: ActionType::Copy,
-			destination: Some("C:/nonexistent_root_zzz_99999/{Date}".to_string()),
-			overwrite: Some(true),
-			verify_integrity: Some(false),
-			preserve_structure: Some(false),
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_static_root_of_destination_no_placeholder() {
-		assert_eq!(static_root_of_destination("C:/data/backup"), "C:/data/backup");
-	}
-
-	#[test]
-	fn test_static_root_of_destination_with_placeholder() {
-		assert_eq!(static_root_of_destination("C:/data/backup/{Date}"), "C:/data/backup/");
-		assert_eq!(static_root_of_destination("C:/data/backup/{Date}/sub"), "C:/data/backup/");
-	}
-
-	#[test]
-	fn test_static_root_of_destination_backslash_separator() {
-		assert_eq!(static_root_of_destination(r"C:\data\backup\{Date}"), r"C:\data\backup\");
-	}
-
-	#[test]
-	fn test_static_root_of_destination_placeholder_at_start() {
-		assert_eq!(static_root_of_destination("{Date}/sub"), "");
-	}
-
-	#[test]
-	fn test_validate_action_move_valid() {
-		let dest = tempdir().unwrap();
-		let action = ActionConfig {
-			type_: ActionType::Move,
-			destination: Some(dest.path().to_str().unwrap().to_string()),
-			overwrite: Some(false),
-			verify_integrity: Some(false),
-			preserve_structure: Some(true),
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_ok());
-	}
-
-	// =========================================================
-	// validate_action: Command
-	// =========================================================
-
-	#[test]
-	fn test_validate_action_command_valid() {
-		let action = ActionConfig {
-			type_: ActionType::Command,
-			destination: None,
-			overwrite: None,
-			verify_integrity: None,
-			preserve_structure: None,
-			working_dir: Some("".to_string()),
-			shell: Some("cmd".to_string()),
-			command: Some("echo hello".to_string()),
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_ok());
-	}
-
-	#[test]
-	fn test_validate_action_command_missing_shell() {
-		let action = ActionConfig {
-			type_: ActionType::Command,
-			destination: None,
-			overwrite: None,
-			verify_integrity: None,
-			preserve_structure: None,
-			working_dir: Some("".to_string()),
-			shell: None,
-			command: Some("echo hello".to_string()),
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_validate_action_command_missing_command() {
-		let action = ActionConfig {
-			type_: ActionType::Command,
-			destination: None,
-			overwrite: None,
-			verify_integrity: None,
-			preserve_structure: None,
-			working_dir: Some("".to_string()),
-			shell: Some("cmd".to_string()),
-			command: None,
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_validate_action_command_missing_working_dir() {
-		let action = ActionConfig {
-			type_: ActionType::Command,
-			destination: None,
-			overwrite: None,
-			verify_integrity: None,
-			preserve_structure: None,
-			working_dir: None,
-			shell: Some("cmd".to_string()),
-			command: Some("echo hello".to_string()),
-			program: None,
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	// =========================================================
-	// validate_action: Execute
-	// =========================================================
-
-	#[test]
-	fn test_validate_action_execute_valid() {
-		let action = ActionConfig {
-			type_: ActionType::Execute,
-			destination: None,
-			overwrite: None,
-			verify_integrity: None,
-			preserve_structure: None,
-			working_dir: Some("".to_string()),
-			shell: None,
-			command: None,
-			program: Some("notepad.exe".to_string()),
-			args: Some(vec![]),
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_ok());
-	}
-
-	#[test]
-	fn test_validate_action_execute_missing_program() {
-		let action = ActionConfig {
-			type_: ActionType::Execute,
-			destination: None,
-			overwrite: None,
-			verify_integrity: None,
-			preserve_structure: None,
-			working_dir: Some("".to_string()),
-			shell: None,
-			command: None,
-			program: None,
-			args: Some(vec![]),
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_validate_action_execute_missing_args() {
-		let action = ActionConfig {
-			type_: ActionType::Execute,
-			destination: None,
-			overwrite: None,
-			verify_integrity: None,
-			preserve_structure: None,
-			working_dir: Some("".to_string()),
-			shell: None,
-			command: None,
-			program: Some("notepad.exe".to_string()),
-			args: None,
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	#[test]
-	fn test_validate_action_execute_missing_working_dir() {
-		let action = ActionConfig {
-			type_: ActionType::Execute,
-			destination: None,
-			overwrite: None,
-			verify_integrity: None,
-			preserve_structure: None,
-			working_dir: None,
-			shell: None,
-			command: None,
-			program: Some("notepad.exe".to_string()),
-			args: Some(vec!["file.txt".to_string()]),
-			message: None,
-		};
-		assert!(validate_action(&action, "test").is_err());
-	}
-
-	// =========================================================
-	// 正常系: 全体を通したバリデーション
-	// =========================================================
-
-	#[test]
-	fn test_validate_rules_config_valid_copy_rule() {
-		let watch_dir = tempdir().unwrap();
-		let dest_dir = tempdir().unwrap();
-		let toml_str = make_rules_toml(
-			&sanitize_path(watch_dir.path()),
-			&format!(r#"
-				type = "copy"
-				destination = "{}"
-				overwrite = true
-				verify_integrity = false
-				preserve_structure = true
-			"#, sanitize_path(dest_dir.path())),
-		);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_ok());
-	}
-
-	#[test]
-	fn test_validate_rules_config_valid_command_rule() {
-		let watch_dir = tempdir().unwrap();
-		let toml_str = make_rules_toml(
-			&sanitize_path(watch_dir.path()),
-			r#"
-				type = "command"
-				shell = "powershell"
-				command = "Get-Date"
-				working_dir = ""
-			"#,
-		);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		let result = validate_rules_config(&config);
-		assert!(result.is_ok());
-	}
-
-	#[test]
-	fn test_validate_rules_config_multiple_rules() {
-		let watch_dir = tempdir().unwrap();
-		let dest_dir = tempdir().unwrap();
-		let wp = sanitize_path(watch_dir.path());
-		let dp = sanitize_path(dest_dir.path());
-		let toml_str = format!(r#"
-			[[rules]]
-			enabled = true
-			name = "rule-1"
-
-			[rules.watch]
-			path = "{wp}"
-			recursive = true
-			target = "file"
-			include_hidden = false
-			patterns = ["*.csv"]
-			exclude_patterns = []
-			events = ["create"]
-
-			[[rules.actions]]
-			type = "copy"
-			destination = "{dp}"
-			overwrite = true
-			verify_integrity = true
-			preserve_structure = false
-
-			[[rules]]
-			enabled = true
-			name = "rule-2"
-
-			[rules.watch]
-			path = "{wp}"
-			recursive = false
-			target = "directory"
-			include_hidden = true
-			regex = "^backup"
-			exclude_patterns = []
-			events = ["create", "modify"]
-
-			[[rules.actions]]
-			type = "command"
-			shell = "cmd"
-			command = "echo done"
-			working_dir = ""
-		"#);
-		let config: RulesConfig = toml::from_str(&toml_str).unwrap();
-		assert_eq!(config.rules.len(), 2);
-		let result = validate_rules_config(&config);
-		assert!(result.is_ok());
-	}
-
-	// =========================================================
-	// テンプレート TOML パーステスト
-	// Windows バックスラッシュパスが TOML として正しくパースできること
-	// =========================================================
-
-	#[test]
-	fn test_global_template_is_valid_toml() {
-		use crate::templates::GLOBAL_TOML;
-		let result = toml::from_str::<toml::Value>(GLOBAL_TOML);
-		assert!(
-			result.is_ok(),
-			"GLOBAL_TOML テンプレートが TOML としてパースできません: {:?}",
-			result.err()
-		);
-	}
-
-	#[test]
-	fn test_global_template_windows_path_preserved() {
-		use crate::templates::GLOBAL_TOML;
-		let val: toml::Value = toml::from_str(GLOBAL_TOML).unwrap();
-		let dir = val["system_log"]["dir"].as_str().unwrap();
-		assert_eq!(dir, r"C:\logs", "system_log.dir のパスが正しく保持されていません");
-	}
-
-	#[test]
-	fn test_rules_template_is_valid_toml() {
-		use crate::templates::RULES_TOML;
-		let result = toml::from_str::<toml::Value>(RULES_TOML);
-		assert!(
-			result.is_ok(),
-			"RULES_TOML テンプレートが TOML としてパースできません: {:?}",
-			result.err()
-		);
-	}
-
-	#[test]
-	fn test_rules_template_windows_path_preserved() {
-		use crate::templates::RULES_TOML;
-		let val: toml::Value = toml::from_str(RULES_TOML).unwrap();
-		let path = val["rules"][0]["watch"]["path"].as_str().unwrap();
-		assert_eq!(path, r"C:\監視フォルダ", "rules[0].watch.path のパスが正しく保持されていません");
-		let detect_dir = val["rules"][0]["log"]["detect"]["dir"].as_str().unwrap();
-		assert_eq!(detect_dir, r"C:\logs", "rules[0].log.detect.dir のパスが正しく保持されていません");
-		let action_dir = val["rules"][0]["log"]["action"]["dir"].as_str().unwrap();
-		assert_eq!(action_dir, r"C:\logs", "rules[0].log.action.dir のパスが正しく保持されていません");
-	}
-}
+#[path = "config_tests.rs"]
+mod tests;
