+++ cat-watcher/src/router.rs
@@ -429,439 +429,5 @@ pub async fn run_router(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::HashSet;
-    use tempfile::TempDir;
-
-    fn make_rule(watch_path: &str, recursive: bool, patterns: Option<Vec<&str>>) -> CompiledRule {
-        let glob_set = patterns.map(|pats| {
-            let mut builder = GlobSetBuilder::new();
-            for p in pats {
-                builder.add(Glob::new(p).unwrap());
-            }
-            builder.build().unwrap()
-        });
-        CompiledRule {
-            name: format!("rule-{}", watch_path),
-            enabled: true,
-            watch_path: watch_path.to_string(),
-            recursive,
-            target: WatchTarget::Both,
-            include_hidden: false,
-            events: vec![Event::Create],
-            glob_set,
-            exclude_glob_set: None,
-            exclude_regex: None,
-            exclude_dir_glob_set: None,
-            exclude_dir_regex: None,
-            dir_glob_set: None,
-            dir_regex: None,
-            regexes: None,
-            actions: vec![],
-            detect_logger: None,
-            action_logger: None,
-        }
-    }
-
-    fn create_events(e: Event) -> HashSet<Event> {
-        let mut s = HashSet::new();
-        s.insert(e);
-        s
-    }
-
-    // 2つの監視ディレクトリを用意し、片方で検知したファイルが
-    // もう片方のルールに誤マッチしないことを確認する（バグ再現テスト）
-    #[test]
-    fn test_no_cross_directory_match() {
-        let dir_a = TempDir::new().unwrap();
-        let dir_b = TempDir::new().unwrap();
-
-        let rule_a = make_rule(dir_a.path().to_str().unwrap(), false, Some(vec!["*.csv"]));
-        let rule_b = make_rule(dir_b.path().to_str().unwrap(), false, Some(vec!["*.csv"]));
-
-        let file_in_a = dir_a.path().join("data.csv");
-        std::fs::write(&file_in_a, "").unwrap();
-        let events = create_events(Event::Create);
-
-        assert!(evaluate_rule(&file_in_a, &events, None, &rule_a), "dir_a のルールはマッチすべき");
-        assert!(!evaluate_rule(&file_in_a, &events, None, &rule_b), "dir_b のルールはマッチしてはいけない");
-    }
-
-    // recursive=false でサブディレクトリのファイルが除外されることを確認
-    #[test]
-    fn test_non_recursive_excludes_subdir() {
-        let dir = TempDir::new().unwrap();
-        let subdir = dir.path().join("sub");
-        std::fs::create_dir(&subdir).unwrap();
-        let file_in_sub = subdir.join("data.csv");
-        std::fs::write(&file_in_sub, "").unwrap();
-
-        let rule = make_rule(dir.path().to_str().unwrap(), false, Some(vec!["*.csv"]));
-        let events = create_events(Event::Create);
-
-        assert!(!evaluate_rule(&file_in_sub, &events, None, &rule), "サブディレクトリのファイルはマッチしてはいけない");
-    }
-
-    // recursive=true ではサブディレクトリのファイルもマッチすることを確認
-    #[test]
-    fn test_recursive_includes_subdir() {
-        let dir = TempDir::new().unwrap();
-        let subdir = dir.path().join("sub");
-        std::fs::create_dir(&subdir).unwrap();
-        let file_in_sub = subdir.join("data.csv");
-        std::fs::write(&file_in_sub, "").unwrap();
-
-        let rule = make_rule(dir.path().to_str().unwrap(), true, Some(vec!["*.csv"]));
-        let events = create_events(Event::Create);
-
-        assert!(evaluate_rule(&file_in_sub, &events, None, &rule), "recursive=true ならサブディレクトリもマッチすべき");
-    }
-
-    // 直下のファイルは recursive=false でもマッチすることを確認
-    #[test]
-    fn test_non_recursive_matches_direct_child() {
-        let dir = TempDir::new().unwrap();
-        let file = dir.path().join("data.csv");
-        std::fs::write(&file, "").unwrap();
-
-        let rule = make_rule(dir.path().to_str().unwrap(), false, Some(vec!["*.csv"]));
-        let events = create_events(Event::Create);
-
-        assert!(evaluate_rule(&file, &events, None, &rule));
-    }
-
-    // パターンに合わないファイルは除外されることを確認
-    #[test]
-    fn test_pattern_mismatch_excluded() {
-        let dir = TempDir::new().unwrap();
-        let file = dir.path().join("image.png");
-        std::fs::write(&file, "").unwrap();
-
-        let rule = make_rule(dir.path().to_str().unwrap(), false, Some(vec!["*.csv"]));
-        let events = create_events(Event::Create);
-
-        assert!(!evaluate_rule(&file, &events, None, &rule));
-    }
-
-    // -----------------------------------------------------------------
-    // to_config_event: Rename / Modify サブタイプのマッピング (#30 回帰)
-    // -----------------------------------------------------------------
-    use notify::event::{DataChange, RenameMode};
-
-    #[test]
-    fn test_to_config_event_rename_from() {
-        let kind = EventKind::Modify(ModifyKind::Name(RenameMode::From));
-        assert_eq!(to_config_event(&kind), Some(Event::Rename));
-    }
-
-    #[test]
-    fn test_to_config_event_rename_to() {
-        let kind = EventKind::Modify(ModifyKind::Name(RenameMode::To));
-        assert_eq!(to_config_event(&kind), Some(Event::Rename));
-    }
-
-    #[test]
-    fn test_to_config_event_rename_any() {
-        let kind = EventKind::Modify(ModifyKind::Name(RenameMode::Any));
-        assert_eq!(to_config_event(&kind), Some(Event::Rename));
-    }
-
-    #[test]
-    fn test_to_config_event_modify_data_still_modify() {
-        let kind = EventKind::Modify(ModifyKind::Data(DataChange::Content));
-        assert_eq!(to_config_event(&kind), Some(Event::Modify));
-    }
-
-    // -----------------------------------------------------------------
-    // matches_target: kind ベース判定 (#23 / #24 回帰)
-    // -----------------------------------------------------------------
-
-    // target=file + Remove(File) → kind=Some(File) でマッチ (削除後でも判定できる)
-    #[test]
-    fn test_matches_target_file_with_kind_file() {
-        let path = Path::new("/nonexistent/will_not_exist.txt");
-        assert!(matches_target(path, &WatchTarget::File, Some(EntryKind::File)));
-        assert!(!matches_target(path, &WatchTarget::File, Some(EntryKind::Dir)));
-    }
-
-    // target=directory + Remove(Folder) → kind=Some(Dir) でマッチ
-    #[test]
-    fn test_matches_target_directory_with_kind_dir() {
-        let path = Path::new("/nonexistent/will_not_exist_dir");
-        assert!(matches_target(path, &WatchTarget::Directory, Some(EntryKind::Dir)));
-        assert!(!matches_target(path, &WatchTarget::Directory, Some(EntryKind::File)));
-    }
-
-    // target=both は kind に関係なく常に true
-    #[test]
-    fn test_matches_target_both_always_true() {
-        let path = Path::new("/nonexistent/whatever");
-        assert!(matches_target(path, &WatchTarget::Both, Some(EntryKind::File)));
-        assert!(matches_target(path, &WatchTarget::Both, Some(EntryKind::Dir)));
-        assert!(matches_target(path, &WatchTarget::Both, None));
-    }
-
-    // kind=None + 実在ファイル → File にマッチ (Modify/Rename パスの fallback)
-    #[test]
-    fn test_matches_target_file_kind_none_existing_file() {
-        let dir = TempDir::new().unwrap();
-        let file = dir.path().join("real.txt");
-        std::fs::write(&file, "").unwrap();
-        assert!(matches_target(&file, &WatchTarget::File, None));
-        assert!(!matches_target(&file, &WatchTarget::Directory, None));
-    }
-
-    // kind=None + パスなし (旧 Windows の Remove(Any) や Rename(From) の旧パス等)
-    // → 判定不能なので target=file / target=directory どちらにも通さない (厳格)。
-    //   旧 OS で Delete を確実に拾いたい場合は target=both を使う運用とする。
-    //   target=both は kind に依存しないので常に通る。
-    #[test]
-    fn test_matches_target_kind_none_path_missing_does_not_match() {
-        let path = Path::new("/definitely/does/not/exist_xyz_12345");
-        assert!(!matches_target(path, &WatchTarget::File, None));
-        assert!(!matches_target(path, &WatchTarget::Directory, None));
-        // target=both は通る
-        assert!(matches_target(path, &WatchTarget::Both, None));
-    }
-
-    // -----------------------------------------------------------------
-    // exclude_regex: ファイル名正規表現除外 (#28)
-    // -----------------------------------------------------------------
-
-    #[test]
-    fn test_exclude_regex_excludes_matching_file() {
-        let dir = TempDir::new().unwrap();
-        let file = dir.path().join("debug_001.log");
-        std::fs::write(&file, "").unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), false, None);
-        rule.exclude_regex = Some(Regex::new(r"^debug_\d+").unwrap());
-        let events = create_events(Event::Create);
-
-        assert!(!evaluate_rule(&file, &events, None, &rule));
-    }
-
-    #[test]
-    fn test_exclude_regex_passes_non_matching_file() {
-        let dir = TempDir::new().unwrap();
-        let file = dir.path().join("report_001.log");
-        std::fs::write(&file, "").unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), false, None);
-        rule.exclude_regex = Some(Regex::new(r"^debug_\d+").unwrap());
-        let events = create_events(Event::Create);
-
-        assert!(evaluate_rule(&file, &events, None, &rule));
-    }
-
-    // -----------------------------------------------------------------
-    // exclude_dir_patterns: フォルダ名 glob 除外 (#28)
-    // -----------------------------------------------------------------
-
-    #[test]
-    fn test_exclude_dir_patterns_excludes_file_in_matching_dir() {
-        let dir = TempDir::new().unwrap();
-        let node_modules = dir.path().join("node_modules");
-        std::fs::create_dir(&node_modules).unwrap();
-        let file = node_modules.join("index.js");
-        std::fs::write(&file, "").unwrap();
-
-        let mut builder = GlobSetBuilder::new();
-        builder.add(Glob::new("node_modules").unwrap());
-        let gs = builder.build().unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.exclude_dir_glob_set = Some(gs);
-        let events = create_events(Event::Create);
-
-        assert!(!evaluate_rule(&file, &events, None, &rule));
-    }
-
-    #[test]
-    fn test_exclude_dir_patterns_passes_file_in_non_matching_dir() {
-        let dir = TempDir::new().unwrap();
-        let src = dir.path().join("src");
-        std::fs::create_dir(&src).unwrap();
-        let file = src.join("main.rs");
-        std::fs::write(&file, "").unwrap();
-
-        let mut builder = GlobSetBuilder::new();
-        builder.add(Glob::new("node_modules").unwrap());
-        let gs = builder.build().unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.exclude_dir_glob_set = Some(gs);
-        let events = create_events(Event::Create);
-
-        assert!(evaluate_rule(&file, &events, None, &rule));
-    }
-
-    #[test]
-    fn test_exclude_dir_patterns_nested_dir_excluded() {
-        let dir = TempDir::new().unwrap();
-        let parent = dir.path().join("packages");
-        let node_modules = parent.join("node_modules");
-        std::fs::create_dir_all(&node_modules).unwrap();
-        let file = node_modules.join("dep.js");
-        std::fs::write(&file, "").unwrap();
-
-        let mut builder = GlobSetBuilder::new();
-        builder.add(Glob::new("node_modules").unwrap());
-        let gs = builder.build().unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.exclude_dir_glob_set = Some(gs);
-        let events = create_events(Event::Create);
-
-        assert!(!evaluate_rule(&file, &events, None, &rule));
-    }
-
-    // -----------------------------------------------------------------
-    // exclude_dir_regex: フォルダ名正規表現除外 (#28)
-    // -----------------------------------------------------------------
-
-    #[test]
-    fn test_exclude_dir_regex_excludes_file_in_matching_dir() {
-        let dir = TempDir::new().unwrap();
-        let hidden = dir.path().join(".cache");
-        std::fs::create_dir(&hidden).unwrap();
-        let file = hidden.join("data.bin");
-        std::fs::write(&file, "").unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.exclude_dir_regex = Some(Regex::new(r"^\.").unwrap());
-        let events = create_events(Event::Create);
-
-        assert!(!evaluate_rule(&file, &events, None, &rule));
-    }
-
-    #[test]
-    fn test_exclude_dir_regex_passes_file_in_non_matching_dir() {
-        let dir = TempDir::new().unwrap();
-        let visible = dir.path().join("data");
-        std::fs::create_dir(&visible).unwrap();
-        let file = visible.join("report.csv");
-        std::fs::write(&file, "").unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.exclude_dir_regex = Some(Regex::new(r"^\.").unwrap());
-        let events = create_events(Event::Create);
-
-        assert!(evaluate_rule(&file, &events, None, &rule));
-    }
-
-    // -----------------------------------------------------------------
-    // dir_patterns: フォルダ名 glob 包含 (#28)
-    // -----------------------------------------------------------------
-
-    #[test]
-    fn test_dir_patterns_includes_file_in_matching_dir() {
-        let dir = TempDir::new().unwrap();
-        let src = dir.path().join("src");
-        std::fs::create_dir(&src).unwrap();
-        let file = src.join("main.rs");
-        std::fs::write(&file, "").unwrap();
-
-        let mut builder = GlobSetBuilder::new();
-        builder.add(Glob::new("src").unwrap());
-        let gs = builder.build().unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.dir_glob_set = Some(gs);
-        let events = create_events(Event::Create);
-
-        assert!(evaluate_rule(&file, &events, None, &rule));
-    }
-
-    #[test]
-    fn test_dir_patterns_excludes_file_in_non_matching_dir() {
-        let dir = TempDir::new().unwrap();
-        let lib = dir.path().join("lib");
-        std::fs::create_dir(&lib).unwrap();
-        let file = lib.join("utils.rs");
-        std::fs::write(&file, "").unwrap();
-
-        let mut builder = GlobSetBuilder::new();
-        builder.add(Glob::new("src").unwrap());
-        let gs = builder.build().unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.dir_glob_set = Some(gs);
-        let events = create_events(Event::Create);
-
-        assert!(!evaluate_rule(&file, &events, None, &rule));
-    }
-
-    #[test]
-    fn test_dir_patterns_direct_child_not_matched() {
-        // watch 直下のファイルはディレクトリコンポーネントが存在しないためマッチしない
-        let dir = TempDir::new().unwrap();
-        let file = dir.path().join("root.csv");
-        std::fs::write(&file, "").unwrap();
-
-        let mut builder = GlobSetBuilder::new();
-        builder.add(Glob::new("src").unwrap());
-        let gs = builder.build().unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), false, None);
-        rule.dir_glob_set = Some(gs);
-        let events = create_events(Event::Create);
-
-        assert!(!evaluate_rule(&file, &events, None, &rule));
-    }
-
-    #[test]
-    fn test_dir_patterns_nested_dir_matched() {
-        // 深いネストでも途中にマッチするフォルダがあれば通る
-        let dir = TempDir::new().unwrap();
-        let nested = dir.path().join("packages").join("src").join("components");
-        std::fs::create_dir_all(&nested).unwrap();
-        let file = nested.join("button.tsx");
-        std::fs::write(&file, "").unwrap();
-
-        let mut builder = GlobSetBuilder::new();
-        builder.add(Glob::new("src").unwrap());
-        let gs = builder.build().unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.dir_glob_set = Some(gs);
-        let events = create_events(Event::Create);
-
-        assert!(evaluate_rule(&file, &events, None, &rule));
-    }
-
-    // -----------------------------------------------------------------
-    // dir_regex: フォルダ名正規表現包含 (#28)
-    // -----------------------------------------------------------------
-
-    #[test]
-    fn test_dir_regex_includes_file_in_matching_dir() {
-        let dir = TempDir::new().unwrap();
-        let reports = dir.path().join("reports_2024");
-        std::fs::create_dir(&reports).unwrap();
-        let file = reports.join("data.csv");
-        std::fs::write(&file, "").unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.dir_regex = Some(Regex::new(r"^reports_\d+$").unwrap());
-        let events = create_events(Event::Create);
-
-        assert!(evaluate_rule(&file, &events, None, &rule));
-    }
-
-    #[test]
-    fn test_dir_regex_excludes_file_in_non_matching_dir() {
-        let dir = TempDir::new().unwrap();
-        let other = dir.path().join("archives");
-        std::fs::create_dir(&other).unwrap();
-        let file = other.join("data.csv");
-        std::fs::write(&file, "").unwrap();
-
-        let mut rule = make_rule(dir.path().to_str().unwrap(), true, None);
-        rule.dir_regex = Some(Regex::new(r"^reports_\d+$").unwrap());
-        let events = create_events(Event::Create);
-
-        assert!(!evaluate_rule(&file, &events, None, &rule));
-    }
-}
+#[path = "router_tests.rs"]
+mod tests;
